Division hashing and list appending collision
I used the division hashing technique w/ the list appending collision method. I went with this because I assumed it would limit the amount of collisions.
The statistics are:
Movie had 14003 collisions, 0 unused spaces and a construction time of .14 seconds.
Quote had 14003 collisions, 0 unused spaces and a construction time of .17 seconds.